requirements.txt
streamlit
opencv-python
zxing
requests
numpy
pillow
